+++ com.unity.render-pipelines.high-definition/Editor/ShaderGraph/HDSubShaderUtilities.cs
@@ -1321,6 +1321,11 @@ namespace UnityEditor.Rendering.HighDefinition
             });
         }
 
+        public static void AddRayTracingProperty(PropertyCollector collector, bool isRayTracing)
+        {
+            collector.AddToggleProperty("_RayTracing", isRayTracing);
+        }
+        
         public static string RenderQueueName(HDRenderQueue.RenderQueueType value)
         {
             switch (value)
@@ -1338,18 +1343,6 @@ namespace UnityEditor.Rendering.HighDefinition
                 case HDRenderQueue.RenderQueueType.AfterPostprocessTransparent:
                     return "After Post-process";
 
-                case HDRenderQueue.RenderQueueType.RaytracingOpaque:
-                {
-                    if ((RenderPipelineManager.currentPipeline as HDRenderPipeline).rayTracingSupported)
-                        return "RayTracing";
-                    return "None";
-                }
-                case HDRenderQueue.RenderQueueType.RaytracingTransparent:
-                {
-                    if ((RenderPipelineManager.currentPipeline as HDRenderPipeline).rayTracingSupported)
-                        return "RayTracing";
-                    return "None";
-                }
                 default:
                     return "None";
             }
@@ -1358,15 +1351,12 @@ namespace UnityEditor.Rendering.HighDefinition
         public static System.Collections.Generic.List<HDRenderQueue.RenderQueueType> GetRenderingPassList(bool opaque, bool needAfterPostProcess)
         {
             // We can't use RenderPipelineManager.currentPipeline here because this is called before HDRP is created by SG window
-            bool supportsRayTracing = HDRenderPipeline.GatherRayTracingSupport(HDRenderPipeline.currentAsset.currentPlatformRenderPipelineSettings);
             var result = new System.Collections.Generic.List<HDRenderQueue.RenderQueueType>();
             if (opaque)
             {
                 result.Add(HDRenderQueue.RenderQueueType.Opaque);
                 if (needAfterPostProcess)
                     result.Add(HDRenderQueue.RenderQueueType.AfterPostProcessOpaque);
-                if (supportsRayTracing)
-                    result.Add(HDRenderQueue.RenderQueueType.RaytracingOpaque);
             }
             else
             {
@@ -1375,8 +1365,6 @@ namespace UnityEditor.Rendering.HighDefinition
                 result.Add(HDRenderQueue.RenderQueueType.LowTransparent);
                 if (needAfterPostProcess)
                     result.Add(HDRenderQueue.RenderQueueType.AfterPostprocessTransparent);
-                if (supportsRayTracing)
-                    result.Add(HDRenderQueue.RenderQueueType.RaytracingTransparent);
             }
 
             return result;